requirements.txt
# data science
numpy>=1.26.4
pandas>=2.2.3
scipy>=1.14.1
scikit-learn>=1.5.2
joblib>=1.4.2

# ml models
xgboost>=2.1.2
imbalanced-learn>=0.12.4

# dashboard
streamlit>=1.40.2
plotly>=5.24.1

# file format
openpyxl>=3.1.0
pyarrow>=18.0.0

# llm / ai agent
openai>=1.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0
